% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata to Fragments List}
\usage{
add_metadata(
  fragments_list,
  metadata_data.frame,
  unique_id = "sample_file_name",
  plate_id = "plate_id",
  group_id = "group_id",
  metrics_baseline_control = "metrics_baseline_control_TF",
  size_standard = "size_standard",
  size_standard_repeat_length = "size_standard_repeat_length"
)
}
\arguments{
\item{fragments_list}{A list of fragment objects to which metadata will be added.}

\item{metadata_data.frame}{A data frame containing the metadata information.}

\item{unique_id}{A character string indicating the column name for unique sample identifiers in the metadata.}

\item{plate_id}{A character string indicating the column name for plate identifiers in the metadata.}

\item{group_id}{A character string indicating the column name for sample group identifiers in the metadata.}

\item{metrics_baseline_control}{A character string indicating the column name for baseline control indicators in the metadata.}

\item{size_standard}{A character string indicating the column name for repeat positive control indicators in the metadata.}

\item{size_standard_repeat_length}{A character string indicating the column name for repeat positive control lengths in the metadata.}
}
\value{
A modified list of fragment objects with added metadata.
}
\description{
This function adds metadata information to a list of fragments.
}
\details{
This function adds specified metadata attributes to each fragment in the list.
It matches the unique sample identifiers from the fragments list with those in the metadata data frame.
}
\examples{

gm_raw <- instability::example_data
metadata <- instability::metadata

test_fragments <- peak_table_to_fragments(gm_raw,
  data_format = "genemapper5",
  dye_channel = "B")

test_metadata <- add_metadata(
  fragments_list = test_fragments,
  metadata_data.frame = metadata,
  unique_id = "unique_id",
  plate_id = "plate_id",
  group_id = "cell_line",
  metrics_baseline_control = "metrics_baseline_control_TF",
  size_standard = "repeat_positive_control_TF",
  size_standard_repeat_length = "repeat_positive_control_length")
}
