% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{call_repeats}
\alias{call_repeats}
\title{Call Repeats for Fragments}
\usage{
call_repeats(
  fragments_list,
  repeat_algorithm = "simple",
  assay_size_without_repeat = 87,
  repeat_size = 3,
  repeat_length_correction = "none"
)
}
\arguments{
\item{fragments_list}{A list of bp_fragments objects containing fragment data.}

\item{repeat_algorithm}{A character specifying the repeat calling algorithm. Options: \code{"simple"} or \code{"nearest_peak"}.}

\item{assay_size_without_repeat}{An integer specifying the assay size without repeat for repeat calling. Default is 87.}

\item{repeat_size}{An integer specifying the repeat size for repeat calling. Default is 3.}

\item{repeat_length_correction}{A character specifying the repeat length correction method. Options: \code{"none"}, \code{"from_metadata"}, \code{"from_genemapper"}. Default is \code{"none"}.}
}
\value{
A list of \code{\link{repeats_fragments}} objects with repeat data added.

#' @details
The calculated repeat lengths are assigned to the corresponding peaks in the provided \code{bp_fragments} object. The repeat lengths can be used for downstream instability analysis.

The \code{simple} algorithm is just the repeat size calculated either directly, or when size standards are used to correct the repeat, it's the repeat length calculated from the model of bp vs repeat length.

The \code{nearest_peak} algorithm aims to model and correct for the systematic variation in fragment sizes that occurs over base pairs. It calculates repeat lengths in a way that helps align peaks with the underlying repeat pattern, making the estimation of repeat lengths more reliable relative to the main peak. The calculated repeat lengths start from the main peak's repeat length (repeat length of the main peak calculated with simple algorithm described above) and increase in increments of the specified \code{repeat_size}. This approach is particularly useful for mitigating the impact of size measurement underestimate, often referred to as "drift," that can occur over base pairs. By ensuring that the calculated repeat lengths are evenly spaced apart by a fixed amount (\code{repeat_size}), this algorithm helps stabilize the estimation of repeat lengths across peaks, leading to more consistent results.
}
\description{
This function calls the repeat lengths for a list of fragments.
}
\seealso{
\code{\link{find_main_peaks}}
}
