% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract_alleles}
\alias{extract_alleles}
\title{Extract Modal Peaks}
\usage{
extract_alleles(fragments_list)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}
}
\value{
A data.frame containing modal peak information for each sample.
}
\description{
Extracts modal peak information from each sample in a list of fragments.
}
