% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract_fragments}
\alias{extract_fragments}
\title{Extract All Fragments}
\usage{
extract_fragments(fragments_list)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}
}
\value{
A data.frame containing peak data for each sample.
}
\description{
Extracts peak data from each sample in a list of fragments.
}
\examples{
 gm_raw <- instability::example_data
 metadata <- instability::metadata

 test_fragments <- peak_table_to_fragments(gm_raw,
   data_format = "genemapper5",
   dye_channel = "B",
   min_size_bp = 400)

 test_metadata <- add_metadata(
   fragments_list = test_fragments,
   metadata_data.frame = metadata,
   unique_id = "unique_id",
   plate_id = "plate_id",
   sample_group_id = "cell_line",
   metrics_baseline_control = "metrics_baseline_control_TF",
   repeat_positive_control_TF = "repeat_positive_control_TF",
   repeat_positive_control_length = "repeat_positive_control_length")

 test_alleles <- find_alleles(
   fragments_list = test_metadata,
   number_of_peaks_to_return = 1,
   peak_region_size_gap_threshold = 6,
   peak_region_height_threshold_multiplier = 1)

 test_repeats <- call_repeats(
   fragments_list = test_alleles,
   repeat_algorithm = "simple",
   assay_size_without_repeat = 87,
   repeat_size = 3,
   repeat_length_correction = "none"
  )

  extract_alleles(test_repeats)

}
