% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsa.R
\name{find_fragment_peaks}
\alias{find_fragment_peaks}
\title{Title}
\usage{
find_fragment_peaks(
  fragments_trace,
  smoothing_window,
  minumum_peak_signal,
  min_bp_size,
  max_bp_size
)
}
\arguments{
\item{max_bp_size}{}
}
\description{
Title
}
