% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{find_fragments}
\alias{find_fragments}
\title{Title}
\usage{
find_fragments(
  fragments_trace_list,
  smoothing_window = 5,
  minumum_peak_signal = 20,
  min_bp_size = 100,
  max_bp_size = 1000
)
}
\arguments{
\item{max_bp_size}{}
}
\description{
Title
}
