% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{find_ladders}
\alias{find_ladders}
\title{Find the ladder peaks in the fsa file and use that to call bp size}
\usage{
find_ladders(
  fsa_list,
  ladder_channel = "DATA.105",
  signal_channel = "DATA.1",
  ladder_sizes = NULL,
  hq_ladder = TRUE,
  spike_location = NULL,
  smoothing_window = 5,
  max_combinations = 2500000,
  ladder_selection_window = 5
)
}
\arguments{
\item{fsa_list}{list from 'read_fsa' function}

\item{ladder_channel}{string: which channel in the fsa file contains the
ladder signal}

\item{signal_channel}{string: which channel in the fsa file contains the data
signal}

\item{ladder_sizes}{numeric vector: bp sizes of ladder used in fragment analysis.
defaults to GeneScan™ 500 LIZ™}

\item{hq_ladder}{logical: If TRUE, c(35, 250, 340) will be dropped from ladder}

\item{spike_location}{numeric: indicate the scan number of the large spike peak}

\item{smoothing_window}{numeric: scan window size for smoothing ladder signal}

\item{max_combinations}{numeric: what is the maximum number of ladder
combinations that should be tested}

\item{ladder_selection_window}{numeric: in the ladder assigning algorithm,
the we iterate through the scans in blocks and test their linear fit.
In fragment analysis, we can assume that the ladder is linear over a
short distance. This value defines how large that block of peaks should be.}
}
\value{
list of fragments_trace objects
}
\description{
Find the ladder peaks in the fsa file and use that to call bp size
}
