% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{fix_ladders}
\alias{fix_ladders}
\title{Title}
\usage{
fix_ladders(
  fragments_trace_list,
  unique_ids,
  size_threshold = 60,
  size_tolerance = 2.5,
  rsq_threshold = 0.9985
)
}
\arguments{
\item{rsq_threshold}{}
}
\description{
Title
}
