% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{fix_ladders_auto}
\alias{fix_ladders_auto}
\title{Automatically fix ladders}
\usage{
fix_ladders_auto(
  fragments_trace_list,
  unique_ids,
  size_threshold = 60,
  size_tolerance = 2.5,
  rsq_threshold = 0.9985
)
}
\arguments{
\item{fragments_trace_list}{list of fragments_trace objects. Output from find_ladders()}

\item{unique_ids}{a character vector of unique ids of samples to fix}

\item{size_threshold}{the bp size from the edges of the size standard to look.
Increase to try and fix ladder peaks further away from correctly assigned.}

\item{size_tolerance}{bp size for how far away prediction is from a real peak
to be counted as correct.}

\item{rsq_threshold}{R-squared threshold of linear models that represent assumed
correct peak assignment.}
}
\value{
a list of fragments_trace object, equal length and names to the list input
}
\description{
This function tries to automatically identify badly assigned peaks and reassign
them. It will only work if some consecutive peaks are correctly assigned.
}
\details{
This function takes a list of fragments_traces objects and tries to automatically
fix the ladder. It uses an assumption that some of the peaks are correct (high
r-squared of three size standard linear model) and then uses those to predict
where the size standard should be. If the size standard is where it's expected,
it will be assigned, otherwise it will be dropped.

The ladders should be manually inspected to confirm correct assignment.
}
\examples{


file_list <- instability::cell_line_fsa_list

  test_ladders <- find_ladders(file_list,
                               ladder_channel = "DATA.105",
                               signal_channel = "DATA.1",
                               ladder_sizes = c(35, 50, 75, 100, 139, 150, 160, 200, 250, 300, 340, 350, 400, 450, 490, 500),
                               hq_ladder = FALSE,
                               max_combinations = 2500000,
                               ladder_selection_window = 8)

  ladders_fixed_list <- fix_ladders_auto(test_ladders, "20230413_B03.fsa")



}
\seealso{
\code{\link[=fix_ladders_manual]{fix_ladders_manual()}} to fix ladders with incorrectly assigned peaks
manually. \code{\link[=plot_ladders]{plot_ladders()}} to plot the assigned ladder peaks onto the raw ladder
signal.
}
