% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{generate_instability_template}
\alias{generate_instability_template}
\title{Generate a Quarto file that has the instability pipeline preset}
\usage{
generate_instability_template(file_name = NULL)
}
\arguments{
\item{file_name}{Name of file to create}
}
\value{
A Quarto template file
}
\description{
Generate a Quarto file that has the instability pipeline preset
}
\examples{

if (interactive()) {
  generate_instability_template("test")
}


}
