% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{instability_quarto_template}
\alias{instability_quarto_template}
\title{instability_quarto_template}
\usage{
instability_quarto_template(file_name = NULL, template = "mouse")
}
\arguments{
\item{file_name}{Name of file to create}

\item{template}{Type of template to produce. Chose from: "mouse"}
}
\value{
A Quarto template file
}
\description{
instability_quarto_template
}
\examples{

#instability_quarto_template("test", "mouse")

}
