% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{plot_repeat_correction_model}
\alias{plot_repeat_correction_model}
\title{Plot Repeat Correction Model}
\usage{
plot_repeat_correction_model(fragments_list)
}
\arguments{
\item{fragments_list}{A list of repeats_fragments class objects obtained from the 'call_repeats' function when the 'repeat_length_correction' was either 'from_metadata' or 'from_genemapper'.}
}
\value{
A ggplot object displaying the repeat correction model results.
}
\description{
Plots the results of the repeat correction model for a list of fragments.
}
