% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{plot_traces}
\alias{plot_traces}
\title{Title}
\usage{
plot_traces(
  fragments_list,
  show_peaks = TRUE,
  n_facet_col = 2,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{ylim}{}
}
\description{
Title
}
