% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{read_fsa}
\alias{read_fsa}
\title{Read fsa file}
\usage{
read_fsa(files)
}
\arguments{
\item{files}{folder path containing the fsa files}
}
\description{
This function is just a wrapper of seqinr::read.abif where each samples is an
element of a list
}
