% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{remove_fragments}
\alias{remove_fragments}
\title{Remove Samples from List}
\usage{
remove_fragments(fragments_list, samples_to_remove)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}

\item{samples_to_remove}{A character vector containing the unique IDs of the samples to be removed.}
}
\value{
A modified list of fragments with the specified samples removed.
}
\description{
A convenient function to remove specific samples from a list of fragments.
}
