% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{repeat_table_to_repeats}
\alias{repeat_table_to_repeats}
\title{Convert Repeat Table to Repeats Fragments}
\usage{
repeat_table_to_repeats(
  df,
  repeat_class = "repeats_fragments",
  repeat_col,
  frequency_col,
  unique_id
)
}
\arguments{
\item{df}{A data frame containing the repeat data.}

\item{repeat_col}{A character string indicating the column name for the repeats.}

\item{frequency_col}{A character string indicating the column name for the repeat frequencies.}

\item{unique_id}{A character string indicating the column name for unique identifiers.}
}
\value{
A list of repeats fragments.
}
\description{
This function converts a repeat table data frame into a list of repeats fragments class.
}
\details{
This function takes a repeat table data frame and converts it into a list of repeats fragments.
The function allows specifying column names for repeats, frequencies, and unique identifiers.
}
