% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extract_fragments}
\alias{extract_fragments}
\title{Extract All Fragments}
\usage{
extract_fragments(fragments_list)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}
}
\value{
A data.frame containing peak data for each sample.
}
\description{
Extracts peak data from each sample in a list of fragments.
}
