% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{plot_fragments}
\alias{plot_fragments}
\title{Plot Peak Data}
\usage{
plot_fragments(
  fragments_list,
  sample_subset = NULL,
  zoom_in_on_peak = NA,
  zoom_range = NULL,
  show_peak_regions = FALSE,
  col_width = NULL,
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_scales = "fixed"
)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}

\item{sample_subset}{A character vector of unique sample IDs of samples to pick out and plot.}

\item{zoom_in_on_peak}{A numeric value (1 or 2) for which allele's peak should be zoomed in on. Use NA to show the whole trace.}

\item{zoom_range}{A numeric vector with a length of 2 to indicate the size range around the zoomed-in peak. For example, c(-20, 20).}

\item{show_peak_regions}{A logical value indicating whether peak regions should be included in the plot.}

\item{col_width}{A numeric value for the column widths of the plotted bars.}

\item{facet_nrow}{A numeric value indicating the number of rows for faceting in the plot (passed to 'nrow' in \code{facet_wrap()}).}

\item{facet_ncol}{A numeric value indicating the number of columns for faceting in the plot (passed to 'ncol' in \code{facet_wrap()}).}

\item{facet_scales}{A character string for setting the axis scales for faceting (passed to 'scales' in \code{facet_wrap()}).}
}
\value{
A ggplot object displaying the peak data.
}
\description{
Plots peak data from a list of fragments.
}
