% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{remove_fragments}
\alias{remove_fragments}
\title{Remove Samples from List}
\usage{
remove_fragments(fragments_list, samples_to_remove)
}
\arguments{
\item{fragments_list}{A list of "fragments" class objects (can be either "bp_fragments" or "repeats_fragments" class).}

\item{samples_to_remove}{A character vector containing the unique IDs of the samples to be removed.}
}
\value{
A modified list of fragments with the specified samples removed.
}
\description{
A convenient function to remove specific samples from a list of fragments.
}
\examples{
gm_raw <- instability::example_data
metadata <- instability::metadata

test_fragments <- peak_table_to_fragments(
  gm_raw,
  data_format = "genemapper5",
  dye_channel = "B")

all_fragment_names <- names(test_fragments)

#pull out unique ids of samples to remove
samples_to_remove <- all_fragment_names[c(1,5,10)]

samples_removed <- remove_fragments(test_fragments, samples_to_remove)

}
